Rename all files from .js to .ts
