+++ src/dry-resources.ts
@@ -1,4 +1,4 @@
-import makeDomStatic from './make-dom-static/index.js'
+import makeDomStatic from './make-dom-static/index.ts'
 
 /**
  * "Dry" the resource+subresources to make them static and context-free.
