--- src/extract-links/index.js
@@ -1,2 +1,0 @@
-export { extractLinksFromCss, extractLinksFromCssSynced } from './from-css.js'
-export { extractLinksFromDom } from './from-dom.js'
