+++ src/extract-links/index.ts
@@ -1,0 +1,2 @@
+export { extractLinksFromCss, extractLinksFromCssSynced } from './from-css.ts'
+export { extractLinksFromDom } from './from-dom.ts'
