+++ src/extract-links/parse-tools.ts
@@ -2,7 +2,7 @@
 // Intended to, for example, create a live view on the URLs inside an element's style or srcset
 // attribute, that allows both reading and writing the URLs in place.
 
-import { memoize, mutableProxyFactory } from '../package.js'
+import { memoize, mutableProxyFactory } from '../package.ts'
 
 /**
  * Allows manipulating tokens within a string.
