+++ src/extract-links/url-attributes/split-token-list.ts
