+++ src/make-dom-static/remove-scripts.ts
