+++ src/set-content-security-policy/index.ts
