+++ src/set-memento-tags.test.ts
@@ -1,4 +1,4 @@
-import { datetimeToString } from './set-memento-tags.js'
+import { datetimeToString } from './set-memento-tags.ts'
 
 describe('datetimeToString', () => {
     test('should work for a basic example', () => {
